Restore TreeVisitor
