+++ javaparser-core/src/main/java/com/github/javaparser/ast/visitor/TreeVisitor.java
@@ -23,7 +23,9 @@ package com.github.javaparser.ast.visitor;
 
 import com.github.javaparser.ast.Node;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.Queue;
 
 /**
  * Iterate over all the nodes in (a part of) the AST.
@@ -31,178 +33,47 @@ import java.util.*;
 public abstract class TreeVisitor {
 
     public void visitLeavesFirst(Node node) {
-        visitPostOrder(node);
+        for (Node child : node.getChildNodes()) {
+            visitLeavesFirst(child);
+        }
+        process(node);
     }
 
     /**
-     * Performs a pre-order node traversal starting with a given node. When each node is visited, {@link #process(Node)}
-     * is called for further processing.
+     * Performs a pre-order node traversal starting with a given node. When each node is visited,
+     * {@link #process(Node)} is called for further processing.
      *
      * @param node The node at which the traversal begins.
+     *
      * @see <a href="https://en.wikipedia.org/wiki/Pre-order">Pre-order traversal</a>
      */
     public void visitPreOrder(Node node) {
-        new PreOrderIterator(node).forEachRemaining(this::process);
-    }
-
-    public static class BreadthFirstIterator implements Iterator<Node> {
-        private final Queue<Node> queue = new LinkedList<>();
-
-        public BreadthFirstIterator(Node node) {
-            queue.add(node);
-        }
-
-        @Override
-        public boolean hasNext() {
-            return !queue.isEmpty();
-        }
-
-        @Override
-        public Node next() {
-            Node next = queue.remove();
-            queue.addAll(next.getChildNodes());
-            return next;
-        }
-    }
-
-    public static class DirectChildrenIterator implements Iterator<Node> {
-        private final Iterator<Node> childrenIterator;
-
-        public DirectChildrenIterator(Node node) {
-            childrenIterator = new ArrayList<>(node.getChildNodes()).iterator();
-        }
-
-        @Override
-        public boolean hasNext() {
-            return childrenIterator.hasNext();
-        }
-
-        @Override
-        public Node next() {
-            return childrenIterator.next();
-        }
+        process(node);
+        new ArrayList<>(node.getChildNodes()).forEach(this::visitPreOrder);
     }
 
     /**
-     * Iterates over the parent of the node, then the parent's parent, then the parent's parent's parent, until running
-     * out of parents.
-     */
-    public static class ParentsVisitor implements Iterator<Node> {
-        private Node node;
-
-        public ParentsVisitor(Node node) {
-            this.node = node;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return node.getParentNode().isPresent();
-        }
-
-        @Override
-        public Node next() {
-            node = node.getParentNode().orElse(null);
-            return node;
-        }
-    }
-
-    public static class PreOrderIterator implements Iterator<Node> {
-        private final Stack<Node> stack = new Stack<>();
-
-        public PreOrderIterator(Node node) {
-            stack.add(node);
-        }
-
-        @Override
-        public boolean hasNext() {
-            return !stack.isEmpty();
-        }
-
-        @Override
-        public Node next() {
-            Node next = stack.pop();
-            List<Node> children = next.getChildNodes();
-            for (int i = children.size() - 1; i >= 0; i--) {
-                stack.add(children.get(i));
-            }
-            return next;
-        }
-    }
-
-    public static class PostOrderIterator implements Iterator<Node> {
-        private final Stack<List<Node>> nodesStack = new Stack<>();
-        private final Stack<Integer> cursorStack = new Stack<>();
-        private final Node root;
-        private boolean hasNext = true;
-
-        public PostOrderIterator(Node root) {
-            this.root = root;
-            fillStackToLeaf(root);
-        }
-
-        private void fillStackToLeaf(Node node) {
-            while (true) {
-                List<Node> childNodes = new ArrayList<>(node.getChildNodes());
-                if (childNodes.isEmpty()) {
-                    break;
-                }
-                nodesStack.push(childNodes);
-                cursorStack.push(0);
-                node = childNodes.get(0);
-            }
-        }
-
-        @Override
-        public boolean hasNext() {
-            return hasNext;
-        }
-
-        @Override
-        public Node next() {
-            final List<Node> nodes = nodesStack.peek();
-            final int cursor = cursorStack.peek();
-            final boolean levelHasNext = cursor < nodes.size();
-            if (levelHasNext) {
-                Node node = nodes.get(cursor);
-                fillStackToLeaf(node);
-                return nextFromLevel();
-            } else {
-                nodesStack.pop();
-                cursorStack.pop();
-                hasNext = !nodesStack.empty();
-                if (hasNext) {
-                    return nextFromLevel();
-                }
-                return root;
-            }
-        }
-
-        private Node nextFromLevel() {
-            final List<Node> nodes = nodesStack.peek();
-            final int cursor = cursorStack.pop();
-            cursorStack.push(cursor + 1);
-            return nodes.get(cursor);
-        }
-    }
-
-    /**
-     * Performs a post-order node traversal starting with a given node. When each node is visited, {@link
-     * #process(Node)} is called for further processing.
+     * Performs a post-order node traversal starting with a given node. When each node is visited,
+     * {@link #process(Node)} is called for further processing.
      *
      * @param node The node at which the traversal begins.
+     *
      * @see <a href="https://en.wikipedia.org/wiki/Post-order">Post-order traversal</a>
      */
     public void visitPostOrder(Node node) {
-        new PostOrderIterator(node).forEachRemaining(this::process);
+        new ArrayList<>(node.getChildNodes()).forEach(this::visitPostOrder);
+        process(node);
     }
 
     /**
-     * Performs a pre-order node traversal starting with a given node. When each node is visited, {@link #process(Node)}
-     * is called for further processing.
+     * Performs a pre-order node traversal starting with a given node. When each node is visited,
+     * {@link #process(Node)} is called for further processing.
+     *
+     * @deprecated As of release 3.1.0, replaced by {@link #visitPreOrder(Node)}
      *
      * @param node The node at which the traversal begins.
+     *
      * @see <a href="https://en.wikipedia.org/wiki/Pre-order">Pre-order traversal</a>
-     * @deprecated As of release 3.1.0, replaced by {@link #visitPreOrder(Node)}
      */
     @Deprecated
     public void visitDepthFirst(Node node) {
@@ -215,26 +86,15 @@ public abstract class TreeVisitor {
      * @param node the start node, and the first one that is passed to process(node).
      */
     public void visitBreadthFirst(Node node) {
-        new BreadthFirstIterator(node).forEachRemaining(this::process);
-    }
-
-    /**
-     * Calls "process" for each direct child node of "node"
-     *
-     * @param node the parent node.
-     */
-    public void visitDirectChildren(Node node) {
-        new DirectChildrenIterator(node).forEachRemaining(this::process);
-    }
-
-    /**
-     * Calls "process" for the parent of "node", then the parent's parent, then the parent's parent's parent, until
-     * running out of parents. <br/>Note that "node" itself is not processed.
-     *
-     * @param node the start node.
-     */
-    public void visitParents(Node node) {
-        new ParentsVisitor(node).forEachRemaining(this::process);
+        final Queue<Node> queue = new LinkedList<>();
+        queue.offer(node);
+        while (queue.size() > 0) {
+            final Node head = queue.peek();
+            for (Node child : head.getChildNodes()) {
+                queue.offer(child);
+            }
+            process(queue.poll());
+        }
     }
 
     /**
